fix all signed indirect pushes ( Ldind_Ix, Ldind_Ux)

+++ source2/IL2CPU/Cosmos.IL2CPU.X86/IL/Ldind_I4.cs
@@ -20,7 +20,7 @@ namespace Cosmos.IL2CPU.X86.IL
                 DestinationReg = CPUx86.Registers.EAX,
                 DestinationIsIndirect = true
             };
-			Assembler.Stack.Push(ILOp.Align(4, 4), typeof(long));
+			Assembler.Stack.Push(ILOp.Align(4, 4), typeof(int));
         }
     }
 }
